refactor: 重构壁纸市场界面
